+++ README.md
@@ -21,10 +21,15 @@ python aio.py
 
 ## Setup
 
-For setting up, just plug into your AioHTTP server.
+### aiohttp
+
+For setting up, just plug into your aiohttp server.
 
 ```python
-subscription_server = WebSocketSubscriptionServer(schema)
+from graphql_ws.aiohttp import AiohttpSubscriptionServer
+
+
+subscription_server = AiohttpSubscriptionServer(schema)
 
 async def subscriptions(request):
     ws = web.WebSocketResponse(protocols=('graphql-ws',))
@@ -59,6 +64,44 @@ class Subscription(graphene.ObjectType):
             yield i
             await asyncio.sleep(1.)
         yield up_to
+
+
+schema = graphene.Schema(query=Query, subscription=Subscription)
+```
+
+
+### Gevent
+
+For setting up, just plug into your Gevent server.
+
+```python
+subscription_server = GeventSubscriptionServer(schema)
+app.app_protocol = lambda environ_path_info: 'graphql-ws'
+
+@sockets.route('/subscriptions')
+def echo_socket(ws):
+    subscription_server.handle(ws)
+    return []
+```
+
+And then, plug into a subscribable schema:
+
+```python
+import graphene
+from rx import Observable
+
+
+class Query(graphene.ObjectType):
+    base = graphene.String()
+
+
+class Subscription(graphene.ObjectType):
+    count_seconds = graphene.Float(up_to=graphene.Int())
+
+    async def resolve_count_seconds(root, info, up_to=5):
+        return Observable.interval(1000)\
+                         .map(lambda i: "{0}".format(i))\
+                         .take_while(lambda i: int(i) <= up_to)
 
 
 schema = graphene.Schema(query=Query, subscription=Subscription)
