--- graphql_ws/graphql_aiows.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""Main module."""
