+++ requirements_dev.txt
@@ -9,4 +9,5 @@ PyYAML==3.11
 pytest==2.9.2
 pytest-runner==2.11.1
 aiohttp==2.1.0
+gevent
 graphene>=2.0
